requirements.txt
anthropic>=0.49.0
httpx>=0.27.0
python-dotenv>=1.0.0
